Cache Name accesses and add simple execution time measurement
New local cache, m_names, that maps name's ID to its row in the
Name table. This showed an initial 10x speedup in the execution
of add_name_id() in a 10,000 row test.

I was frustrated with trying to under gprof's output, so I added
a trivial measurement class that accumulates the time spent in a
particular scope. I first used it in add_name_id(). Very, very
simple and much more direct.