Add initial combat statistics
Added structures and function to fill some simple statistics, like total damage,
damage taken, etc. Totally not done but a start.

Updated db_populator to include the final "exit combat" event as part of the combat.

Added function in LogParserTypes to get the ID from an Actor variant - easier to
have it one place close to the definition.

Added some more searches.

Initial work in preparation for a must faster way to query and pretty print event
info in the searcher.